build with multiple configurations

+++ src/main/psxsdk/libsnd/seqread.c
@@ -164,7 +164,7 @@ void _SsContNrpn2(s16, s16, u8);
 void _SsContRpn1(s16, s16, u8);
 void _SsContRpn2(s16, s16, u8);
 void _SsContResetAll(s16 arg0, s16 arg1);
-void _SsContNrpn1(s16 arg0, s16 arg1, s8 arg2);
+void _SsContNrpn1(s16 arg0, s16 arg1, s16 arg2);
 
 void _SsSetControlChange(s16 arg0, s16 arg1, s32 arg2) {
     u32 control;
@@ -326,7 +326,7 @@ typedef void (*SndSsMarkCallbackProc)(short seq_no, short sep_no, short data);
 
 extern SndSsMarkCallbackProc _SsMarkCallback[32][16];
 
-void _SsContNrpn1(s16 arg0, s16 arg1, s8 arg2) {
+void _SsContNrpn1(s16 arg0, s16 arg1, s16 arg2) {
     SndSsMarkCallbackProc temp_v0;
     struct SeqStruct* temp_s0;
     temp_s0 = &_ss_score[arg0][arg1];
